Sprites können gezeichnet werden / Vectoren / Konstruktoren verbessert

+++ src/spacetraveler/SpaceObjectModel.java
@@ -5,14 +5,14 @@ import org.jsfml.system.*;
 public class SpaceObjectModel {
 	private Vector2f v;		// Bewegungsrichtung / Geschwindigkeit
 	private Vector2f e;		// Energievektor
-	private float m;		// Masse
+	private double m;		// Masse
 	
 	
 	
 	private void updateVelocity(){		
-		float v_abs = (float)(2*Math.sqrt(e.x*e.x + e.y*e.y)/m);
+		double v_abs = 2*Math.sqrt(e.x*e.x + e.y*e.y)/m;
 		Vector2f e0 = Vector2f.div(e,(float)(Math.sqrt(e.x*e.x + e.y*e.y)));
-		this.v = Vector2f.mul(e0, v_abs);
+		this.v = Vector2f.mul(e0, (float)v_abs);
 	}
 	
 //	private Vector2f getEnergy(){
@@ -28,7 +28,7 @@ public class SpaceObjectModel {
 		updateVelocity();
 	}
 	
-	public SpaceObjectModel(float m, Vector2f energy){
+	public SpaceObjectModel(double m, Vector2f energy){
 		this.m = m;
 		this.e = energy;
